# MessageOutput: thread-safety and dynamic memory