Very early on I found issues with the serial output of OpenDTU, i.e., garbled messages, where one message would be interrupted with another message. Also, some output seemed to be missing. There was at least discussion about this, see #1130. Very recently, a change was implemented in 43cba675310e0cfda3e325092aed6555a370ed48, which is a big step forward, but does not solve all issues:

* The MessageOutput buffer for the web console is still limited to an arbitrary amount of 500 Bytes.
* Lots of functions defined in the core's `Print.h` and implemented in `Print.cpp`, especially `println(String&)`, still are "vurnerable" to scheduling. Example: Scheduling can happen after printing the contents of a line, but before printing the `\r\n`.
* Not all messages are printed using `println()`, but are stitched together with multiple `print()`s. Those can still be garbled beyond recognition due to scheduling.
* Minor issue: There is a small chance of a buffer overflow in `MessageOutput::write(uint8_t)` because the lock is not protecting the buffer position check.

My proposal with this MR is as follows:
* Every task/context writes to its own buffer (managed by MessageOutput). This prevents messages being garbled at all.
* These buffers are dynamic, such that there is no issue printing lots of text per line.
* Each time a line is completed, and only then, the task's buffered message is printed on the serial console. Additionally, the whole line is moved to a line-buffer for transport to the websocket.
* In the context of the main loop, the buffered lines are moved to the websocket. It is taken care that the websocket queue is not overrun, so all messages actually appear in the web console.
* The line buffer for the web console within MessageOutput is dynamic, such that no text has to be dropped (RAM limits still apply, of course). Lines are retried once the websocket handler got to process its queue.
* It is taken care that memory that can be freed is freed. Also, buffers of deleted tasks are freed in each loop.