@schlimmchen 2023-08-06
I really don't expect this to be a problem even on our memory-constrained ESP32 devices. The buffers aren't that big (lines of log output), so they can occupy gaps in the memory layout, and more importantly, they are freed as soon as possible and they are indeed all freed routinely as long as the main loop is running. Moreover, I took care that each buffer is not copied uneccesarily (it is moved to `_lines` and then moved to a `shared_ptr` for the websocket, not copied). There is a chance that a buffer must grow, in which case the existing data is copied to a new block. There also is the `unordered_map` that keeps a little memory allocated for each living task. That's the theory.

I hate it but have to admit that your point is of course totally valid. This change must not break existing functionality.

Let me point out though, that being unable to have "nice things" like this working solution to the problem of mangled logs because some other component insists on allocating an insanely huge block of consecutive memory (relative to the whole memory) is frustrating. In the long run, data should be processed in reasonable chunks instead. I hope that's achievable. Since I do want prometheus integration myself in the near future, I might look at this myself.

Having said that, I ran a test like this:
* Uploaded a firmware including this change and started logging the serial output using `minicom -C <file>`.
* Used `websocat` from https://github.com/vi/websocat/.
* Started `websocat ws://<hostname>/livedata`.
* Started `websocat ws://<hostname>/console`.
* Used `while true; do wget http://<hostname>/api/prometheus/metrics -O /tmp/metrics.txt &>/dev/null; [ $? -ne 0 ] && echo "failed $(date -Iseconds)"; sleep 0.1; done` to query the prometheus metrics approximately every 100ms.
* (Made sure that the above command would actually log a message on error by compiling a firmware that always throws `std::bad_alloc` in the prometheus handler.)
* Only one inverter (HM-400) is configured to be polled (does that affect the memory size required? AFAIK the block in the prometheus handler is always 40k?).
* Reduce the inverter poll interval to 1 second by configuration.
* Hacked the MQTT publish interval to 1 second (firmware change). MQTT is in use and connected to the broker.
* Threw in a `while true; do mosquitto_pub -h <host> -t 'solar-staging/<serial>/cmd/limit_nonpersistent_relative' -m 90; sleep 10; mosquitto_pub -h <host> -t 'solar-staging/<serial>/cmd/limit_nonpersistent_relative' -m 100; sleep 10; done` for good measure (the MQTT task logs messages in its own context).
* The heap is 42% used at the start of the test. I can actually observe it jumping to a value of 55% to 58% used, then jump back to 42%, at `/info/system` of the web application, when refreshing the page often.
* My browser window was also using the live view websocket most of the time.

I let this run for one and a half hour and observed no `wget` error and the uptime of the OpenDTU actually reached more than 60 minutes, so no reboot due to an uncatched bad allocation. The minicom log file did not show any "out of resources" messages.

Also, this change is running for ~20 hours on my productive board, with no signs of issues. However, that board runs OpenDTU-OnBattery, so it's not perfectly comparable. Still, I did not cause any issues by querying the prometheus API on that node.

What do you think? Was 1 hour long enough (with the intervals so short)? Have I missed something that I should have run in parallel? Is this setup worth documenting for future stress tests?