@tbnobody 2023-08-22
I tested the PR today. The memory usage was enormous. At some times it consumed 20kb more RAM then without this PR. The response of the system info api was also a little bit laggy. The flash consumption was at about 50kb if I remember correctly. 
This was the reason why I didn't include it in todays release.