@schlimmchen 2023-08-24
## Setup

I have an ESP32-WROOM-32U that I used for testing in more detail to address your concers. I build for the `generic_esp32` environment. It polls one inverter every second. It is connected to an MQTT broker and has a valid local time (NTP). The Wi-Fi AP is stopped, i.e., the uptime is greater than 3 minutes. A web browser is connected to the web console while the tests are conducted. An instance of minicom is connected to the serial port.

This is how I inspected the heap usage:
```
rm /tmp/heap_used.txt; for i in $(seq 0 100); do wget http://192.168.16.155/api/system/status -O - | jq '.heap_used' >> /tmp/heap_used.txt; done
cat /tmp/heap_used.txt | sort | head -n1 # min used
cat /tmp/heap_used.txt | sort | tail -n1 # max used
```

This is how I tested the latency in the system info page:
```
wrk -t1 -c1 -d30s --latency http://192.168.16.155/api/system/status
```

This neglects any delays and any lagging that is caused by the client when executing the web interface. This especially includes checking for a new firmware version online, which can take more or less time depending on various factors.

## Flash consumption

Thanks for checking the flash usage again. It's the same for me. Are you okay with spending ~4.2k of flash on this improvement? We can carve out ~1.7k if we switched to a `std::map` rather than a `std::unordered_map`, which might cause slightly worse performance.

## Interesting problem

I noticed that the web console output and the serial output contains less lines per time while doing any of the two tests and the inverter is not polled successfully while the tests are running. This probably means that processing the HTTP response of `api/system/status` starves the Hoymiles library of processing time. This is true for firmware that includes the change and firmware that does *not* include the change. That's plausible since the change does not affect processing the system info JSON response. This might be worth an investigation on its own. Maybe the Hoymiles `loop()` should run in a different task with higher priority or even on CPU0. That will require making sure that the interlocking of the public interface is sufficiently tight, of course.

## Results

### Heap usage

Heap usage when the change is *not* included is between 124120 and 132708 Bytes.
Heap usage when the change is included is between 124940 and 131032 Bytes. Another run: 123256 to 130464 Bytes.

By chance, I managed to capture a moment where the firmware without the change used more heap memory as with the change included.

It seemed very odd to me that this feature would cause 20k of heap memory being hogged at all and especially for a long enough period that one would actually see it in the system info page.

I understand that memory is a scarce resource and that you have been burned before, but realistically, this change will only use memory in the region of a few kilobytes of memory at the very worst. That could happen if a lot of software components with their own `loop()` write a lot of lines (kilobytes of text) to `MessageOutput` *before* `MessageOutput` can clear out those lines and free the respective memory. That does not happen on a regular basis.

### Latency

Without the change:
```
Running 30s test @ http://192.168.16.155/api/system/status
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   444.35ms    7.35ms 477.13ms   87.50%
    Req/Sec     1.83      0.38     2.00     82.81%
  Latency Distribution
     50%  442.18ms
     75%  447.32ms
     90%  452.91ms
     99%  477.13ms
  64 requests in 30.05s, 41.56KB read
Requests/sec:      2.13
Transfer/sec:      1.38KB
```

With the change:
```
Running 30s test @ http://192.168.16.155/api/system/status
  1 threads and 1 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   444.42ms    5.72ms 466.10ms   73.85%
    Req/Sec     1.91      0.29     2.00     90.77%
  Latency Distribution
     50%  442.70ms
     75%  447.05ms
     90%  450.46ms
     99%  466.10ms
  65 requests in 30.04s, 42.21KB read
Requests/sec:      2.16
Transfer/sec:      1.41KB
```

I conclude that your observation regarding the lagging in the system info page might have something to do with the client-side processing of the page in the browser. It seems that this change has no influence on the time it takes to prepare the information required to render the system info page.

That being said, preparing the JSON response for that page might be in need of improvement. 444ms of exclusive processing time is quite a lot. It's probably that much because coming the memory to calculate the heap usage is quite intense.

## Next Steps

What else should be look at to make sure this change does not cause issues elsewhere?